Refactor human user roles
Previously the playbook was written for only one server, so a single
role created all user accounts. Now we have to control who can login
to which servers, so split things into one role for each person.

Ansible has loops and variables which we could use for user
management, but the present approach is probably much easier.